+++ cli/src/main/kotlin/dev/vulnlog/cli/commands/SerialisationTranslator.kt
@@ -61,7 +61,7 @@ class SerialisationTranslator {
     }
 
     private fun VulnlogAnalysisData.toAnalysis(): Analysis {
-        return Analysis(analysedAt, verdict, reasoning)
+        return Analysis(analysedAt, verdict.level, reasoning)
     }
 
     private fun VulnlogTaskData.toTask(): Task? {