+++ dsl/src/main/kotlin/dev/vulnlog/dsl/VlAnalysis.kt
@@ -4,7 +4,7 @@ import java.time.LocalDate
 
 public class AnalysisBuilder(public val reportData: ReportData) {
     public var analysedAt: LocalDate? = null
-    public var verdict: String? = null
+    public var verdict: VerdictSpecification? = null
     public var reasoning: String? = null
 
     public fun build(): TaskBuilder {
@@ -12,7 +12,66 @@ public class AnalysisBuilder(public val reportData: ReportData) {
     }
 }
 
-public data class AnalysisData(val analysedAt: LocalDate, var verdict: String, var reasoning: String)
+public data class AnalysisData(val analysedAt: LocalDate, var verdict: VerdictSpecification, var reasoning: String)
+
+public sealed interface VerdictSpecification {
+    public val level: String
+}
+
+internal data object VerdictCritical : VerdictSpecification {
+    override val level: String = "critical"
+}
+
+/**
+ * Vulnerability analysis resulted in critical impact to software project.
+ *
+ * @since v0.6.0
+ */
+public val critical: VerdictSpecification = VerdictCritical
+
+internal data object VerdictHigh : VerdictSpecification {
+    override val level: String = "high"
+}
+
+/**
+ * Vulnerability analysis revealed high impact on software project.
+ *
+ * @since v0.6.0
+ */
+public val high: VerdictSpecification = VerdictHigh
+
+internal data object VerdictModerate : VerdictSpecification {
+    override val level: String = "moderate"
+}
+
+/**
+ * Vulnerability analysis revealed moderate impact on software project.
+ *
+ * @since v0.6.0
+ */
+public val moderate: VerdictSpecification = VerdictModerate
+
+internal data object VerdictLow : VerdictSpecification {
+    override val level: String = "low"
+}
+
+/**
+ * Vulnerability analysis revealed low impact on software project.
+ *
+ * @since v0.6.0
+ */
+public val low: VerdictSpecification = VerdictLow
+
+public data object VerdictNotAffected : VerdictSpecification {
+    override val level: String = "not affected"
+}
+
+/**
+ * Vulnerability analysis revealed that the vulnerability does not affect the software project.
+ *
+ * @since v0.6.0
+ */
+public val notAffected: VerdictSpecification = VerdictNotAffected
 
 public interface Verdict {
     /**
@@ -20,7 +79,17 @@ public interface Verdict {
      *
      * @since v0.5.0
      */
+    @Deprecated(
+        message = "The method verdict(String) is deprecated. Use the method verdict(VerdictSpecification) instead.",
+    )
     public infix fun verdict(verdict: String): VlAnalyseReasoningStep
+
+    /**
+     * A verdict based on the analysis of the report on the software project.
+     *
+     * @since v0.6.0
+     */
+    public infix fun verdict(verdict: VerdictSpecification): VlAnalyseReasoningStep
 }
 
 public interface VlAnalyseInitStep : Verdict {
@@ -46,19 +115,19 @@ public interface VlAnalyseReasoningStep {
 
 public sealed interface VulnlogAnalysisData {
     public val analysedAt: LocalDate
-    public val verdict: String
+    public val verdict: VerdictSpecification
     public val reasoning: String
 }
 
 public data class VulnlogAnalysisDataImpl(
     override val analysedAt: LocalDate,
-    override val verdict: String,
+    override val verdict: VerdictSpecification,
     override val reasoning: String,
 ) : VulnlogAnalysisData
 
 public object VulnlogAnalysisDataEmpty : VulnlogAnalysisData {
     override val analysedAt: LocalDate = LocalDate.MIN
-    override val verdict: String = ""
+    override val verdict: VerdictSpecification = notAffected
     override val reasoning: String = ""
 
     override fun toString(): String {