+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Introduce verdict types (`notAffected`, `low`, `moderate`, `high` and `critical`) to the DSL. The existing string
+  based verdict is deprecated and will be removed in an upcoming release.
 - More DSL API documentation.
 
 ## [0.5.3] - 2025-03-12