+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Introduce verdict types (`notAffected`, `low`, `moderate`, `high` and `critical`) to the DSL. The existing string
   based verdict is deprecated and will be removed in an upcoming release.
 - More DSL API documentation.
+- Three default reporter (Snyk Open Source, OWASP Dependency Check and Semgrep Supply Chain).
 
 ## [0.5.3] - 2025-03-12